+++ geode-junit/src/main/java/org/apache/geode/codeAnalysis/AnalyzeSerializablesJUnitTestBase.java
@@ -14,17 +14,14 @@
  */
 package org.apache.geode.codeAnalysis;
 
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.fail;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
-import java.io.BufferedReader;
 import java.io.ByteArrayInputStream;
 import java.io.DataInputStream;
 import java.io.Externalizable;
 import java.io.File;
-import java.io.FileReader;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InvalidClassException;
@@ -34,153 +31,59 @@ import java.lang.reflect.Modifier;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashSet;
-import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
 import java.util.Properties;
 import java.util.ServiceLoader;
 import java.util.Set;
 
-import org.junit.AfterClass;
 import org.junit.Before;
-import org.junit.Rule;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
-import org.junit.rules.TestName;
 
 import org.apache.geode.CancelException;
+import org.apache.geode.DataSerializable;
 import org.apache.geode.DataSerializer;
 import org.apache.geode.codeAnalysis.decode.CompiledClass;
 import org.apache.geode.codeAnalysis.decode.CompiledField;
-import org.apache.geode.codeAnalysis.decode.CompiledMethod;
 import org.apache.geode.distributed.ConfigurationProperties;
 import org.apache.geode.distributed.internal.DistributedSystemService;
 import org.apache.geode.distributed.internal.DistributionConfig;
 import org.apache.geode.distributed.internal.DistributionConfigImpl;
-import org.apache.geode.internal.HeapDataOutputStream;
 import org.apache.geode.internal.InternalDataSerializer;
+import org.apache.geode.internal.serialization.BufferDataOutputStream;
+import org.apache.geode.internal.serialization.DataSerializableFixedID;
 import org.apache.geode.internal.serialization.Version;
 import org.apache.geode.pdx.internal.TypeRegistry;
 import org.apache.geode.test.junit.categories.SerializationTest;
-import org.apache.geode.test.junit.rules.ClassAnalysisRule;
 import org.apache.geode.unsafe.internal.sun.reflect.ReflectionFactory;
 
+/**
+ * This subclass of AbstractAnalyzeSerializablesTestBase uses DataSerializer and
+ * InternalDataSerializer. It also performs initialization of the Geode TypeRegistry
+ */
 @Category({SerializationTest.class})
-public abstract class AnalyzeSerializablesJUnitTestBase {
+public abstract class AnalyzeSerializablesJUnitTestBase extends
+    AnalyzeDataSerializablesJUnitTestBase {
+  protected static final String ACTUAL_SERIALIZABLES_DAT = "actualSerializables.dat";
 
-  private static final String NEW_LINE = System.getProperty("line.separator");
-
-  private static final String FAIL_MESSAGE = NEW_LINE + NEW_LINE
-      + "If the class is not persisted or sent over the wire add it to the file " + NEW_LINE + "%s"
-      + NEW_LINE + "Otherwise if this doesn't break backward compatibility, copy the file "
-      + NEW_LINE + "%s to " + NEW_LINE + "%s.";
-  private static final String EXCLUDED_CLASSES_TXT = "excludedClasses.txt";
-  private static final String ACTUAL_DATA_SERIALIZABLES_DAT = "actualDataSerializables.dat";
-  private static final String ACTUAL_SERIALIZABLES_DAT = "actualSerializables.dat";
-  private static final String OPEN_BUGS_TXT = "openBugs.txt";
-
-  /**
-   * all loaded classes
-   */
-  private Map<String, CompiledClass> classes;
-
-  private File expectedDataSerializablesFile;
-  private String expectedSerializablesFileName =
+  protected String expectedSerializablesFileName =
       "sanctioned-" + getModuleName() + "-serializables.txt";
+  protected List<ClassAndVariableDetails> expectedSerializables;
 
-  private List<ClassAndMethodDetails> expectedDataSerializables;
-  private List<ClassAndVariableDetails> expectedSerializables;
-
-  @Rule
-  public TestName testName = new TestName();
-
-  @Rule
-  public ClassAnalysisRule classProvider = new ClassAnalysisRule(getModuleName());
-
-  private void loadExpectedDataSerializables() throws Exception {
-    this.expectedDataSerializablesFile = getResourceAsFile("sanctionedDataSerializables.txt");
-    assertThat(this.expectedDataSerializablesFile).exists().canRead();
-
-    this.expectedDataSerializables =
-        CompiledClassUtils.loadClassesAndMethods(this.expectedDataSerializablesFile);
-  }
-
-  public void loadExpectedSerializables() throws Exception {
-    try (InputStream expectedSerializablesStream =
-        getResourceAsStream(getModuleClass(), expectedSerializablesFileName)) {
-      // the expectedSerializablesStream will be automatically closed when we exit this block
-      this.expectedSerializables =
-          CompiledClassUtils.loadClassesAndVariables(expectedSerializablesStream);
-    }
-  }
-
-  public void findClasses() throws Exception {
-    classes = classProvider.getClasses();
-
-    List<String> excludedClasses = loadExcludedClasses(getResourceAsFile(EXCLUDED_CLASSES_TXT));
-    List<String> openBugs = loadOpenBugs(getResourceAsFile(OPEN_BUGS_TXT));
-
-    excludedClasses.addAll(openBugs);
-    removeExclusions(classes, excludedClasses);
-  }
 
   @Before
   public void setUp() throws Exception {
     TypeRegistry.init();
   }
 
-  @AfterClass
-  public static void afterClass() {
-    ClassAnalysisRule.clearCache();
-  }
-
-  private List<DistributedSystemService> initializeServices() {
-    ServiceLoader<DistributedSystemService> loader =
-        ServiceLoader.load(DistributedSystemService.class);
-    List<DistributedSystemService> services = new ArrayList<>();
-    for (DistributedSystemService service : loader) {
-      services.add(service);
-    }
-    return services;
-  }
-
-
-  /**
-   * Override only this one method in sub-classes
-   */
-  protected abstract String getModuleName();
-
+  @Override
   protected Class getModuleClass() {
+    // subclasses should override this if their sanctioned-serializables.txt file is not
+    // in org.apache.geode.internal
     return InternalDataSerializer.class;
   }
 
-  @Test
-  public void testDataSerializables() throws Exception {
-    // assumeTrue("Ignoring this test when java version is 9 and above",
-    // !SystemUtils.isJavaVersionAtLeast(JavaVersion.JAVA_9));
-    System.out.println(this.testName.getMethodName() + " starting");
-    findClasses();
-    loadExpectedDataSerializables();
-
-    File actualDataSerializablesFile = createEmptyFile(ACTUAL_DATA_SERIALIZABLES_DAT);
-    System.out.println(this.testName.getMethodName() + " actualDataSerializablesFile="
-        + actualDataSerializablesFile.getAbsolutePath());
-
-    List<ClassAndMethods> actualDataSerializables = findToDatasAndFromDatas();
-    CompiledClassUtils.storeClassesAndMethods(actualDataSerializables, actualDataSerializablesFile);
-
-    String diff =
-        CompiledClassUtils
-            .diffSortedClassesAndMethods(this.expectedDataSerializables, actualDataSerializables);
-    if (!diff.isEmpty()) {
-      System.out.println(
-          "++++++++++++++++++++++++++++++testDataSerializables found discrepancies++++++++++++++++++++++++++++++++++++");
-      System.out.println(diff);
-      fail(diff + FAIL_MESSAGE, getSrcPathFor(getResourceAsFile(EXCLUDED_CLASSES_TXT)),
-          actualDataSerializablesFile.getAbsolutePath(),
-          getSrcPathFor(this.expectedDataSerializablesFile));
-    }
-  }
 
   @Test
   public void testSerializables() throws Exception {
@@ -208,66 +111,11 @@ public abstract class AnalyzeSerializablesJUnitTestBase {
   }
 
   @Test
-  public void testExcludedClassesExistAndDoNotDeserialize() throws Exception {
-    List<String> excludedClasses = loadExcludedClasses(getResourceAsFile(EXCLUDED_CLASSES_TXT));
-    Properties properties = new Properties();
-    properties.put(ConfigurationProperties.VALIDATE_SERIALIZABLE_OBJECTS, "true");
-    properties.put(ConfigurationProperties.SERIALIZABLE_OBJECT_FILTER, "!*");
-    DistributionConfig distributionConfig = new DistributionConfigImpl(properties);
-    InternalDataSerializer.initialize(distributionConfig, initializeServices());
-
-    for (String filePath : excludedClasses) {
-      String className = filePath.replaceAll("/", ".");
-      System.out.println("testing class " + className);
-
-      Class excludedClass = Class.forName(className);
-      assertTrue(
-          excludedClass.getName()
-              + " is not Serializable and should be removed from excludedClasses.txt",
-          Serializable.class.isAssignableFrom(excludedClass));
-
-      if (!excludedClass.isEnum()) {
-        final Object excludedInstance;
-        try {
-          excludedInstance = excludedClass.newInstance();
-        } catch (InstantiationException | IllegalAccessException e) {
-          // okay - it's in the excludedClasses.txt file after all
-          // IllegalAccessException means that the constructor is private.
-          continue;
-        }
-        serializeAndDeserializeObject(excludedInstance);
-      }
-    }
-  }
-
-
-  private void serializeAndDeserializeObject(Object object) throws Exception {
-    HeapDataOutputStream outputStream = new HeapDataOutputStream(Version.CURRENT);
-    try {
-      DataSerializer.writeObject(object, outputStream);
-    } catch (IOException e) {
-      // some classes, such as BackupLock, are Serializable because the extend something
-      // like ReentrantLock but we never serialize them & it doesn't work to try to do so
-      System.out.println("Not Serializable: " + object.getClass().getName());
-    }
-    try {
-      DataSerializer
-          .readObject(new DataInputStream(new ByteArrayInputStream(outputStream.toByteArray())));
-      fail("I was able to deserialize " + object.getClass().getName());
-    } catch (InvalidClassException e) {
-      // expected
-    }
-  }
-
-  @Test
   public void testSanctionedClassesExistAndDoDeserialize() throws Exception {
     loadExpectedSerializables();
     Set<String> openBugs = new HashSet<>(loadOpenBugs(getResourceAsFile(OPEN_BUGS_TXT)));
 
-    DistributionConfig distributionConfig = new DistributionConfigImpl(new Properties());
-    distributionConfig.setValidateSerializableObjects(true);
-    distributionConfig.setSerializableObjectFilter("!*");
-    InternalDataSerializer.initialize(distributionConfig, initializeServices());
+    initializeSerializationService();
 
     for (ClassAndVariableDetails details : expectedSerializables) {
       if (openBugs.contains(details.className)) {
@@ -390,109 +238,15 @@ public abstract class AnalyzeSerializablesJUnitTestBase {
     }
   }
 
-  private void serializeAndDeserializeSanctionedObject(Object object) throws Exception {
-    HeapDataOutputStream outputStream = new HeapDataOutputStream(Version.CURRENT);
-    try {
-      DataSerializer.writeObject(object, outputStream);
-    } catch (IOException e) {
-      // some classes, such as BackupLock, are Serializable because the extend something
-      // like ReentrantLock but we never serialize them & it doesn't work to try to do so
-      throw new AssertionError("Not Serializable: " + object.getClass().getName(), e);
-    }
-    try {
-      Object instance = DataSerializer
-          .readObject(new DataInputStream(new ByteArrayInputStream(outputStream.toByteArray())));
-    } catch (CancelException e) {
-      // PDX classes fish for a PDXRegistry and find that there is no cache
-    } catch (InvalidClassException e) {
-      fail("I was unable to deserialize " + object.getClass().getName(), e);
+  public void loadExpectedSerializables() throws Exception {
+    try (InputStream expectedSerializablesStream =
+        getResourceAsStream(getModuleClass(), expectedSerializablesFileName)) {
+      // the expectedSerializablesStream will be automatically closed when we exit this block
+      this.expectedSerializables =
+          CompiledClassUtils.loadClassesAndVariables(expectedSerializablesStream);
     }
   }
 
-  private String getSrcPathFor(File file) {
-    return getSrcPathFor(file, "test");
-  }
-
-  private String getSrcPathFor(File file, String testOrMain) {
-    return file.getAbsolutePath().replace(
-        "build" + File.separator + "resources" + File.separator + "test",
-        "src" + File.separator + testOrMain + File.separator + "resources");
-  }
-
-  private List<String> loadExcludedClasses(File exclusionsFile) throws IOException {
-    List<String> excludedClasses = new LinkedList<>();
-    FileReader fr = new FileReader(exclusionsFile);
-    BufferedReader br = new BufferedReader(fr);
-    try {
-      String line;
-      while ((line = br.readLine()) != null) {
-        line = line.trim();
-        if (!line.isEmpty() && !line.startsWith("#")) {
-          excludedClasses.add(line);
-        }
-      }
-    } finally {
-      fr.close();
-    }
-    return excludedClasses;
-  }
-
-  private List<String> loadOpenBugs(File exclusionsFile) throws IOException {
-    List<String> excludedClasses = new LinkedList<>();
-    FileReader fr = new FileReader(exclusionsFile);
-    BufferedReader br = new BufferedReader(fr);
-    try {
-      String line;
-      // each line should have bug#,full-class-name
-      while ((line = br.readLine()) != null) {
-        line = line.trim();
-        if (!line.isEmpty() && !line.startsWith("#")) {
-          String[] split = line.split(",");
-          if (split.length != 2) {
-            fail("unable to load classes due to malformed line in openBugs.txt: " + line);
-          }
-          excludedClasses.add(line.split(",")[1].trim());
-        }
-      }
-    } finally {
-      fr.close();
-    }
-    return excludedClasses;
-  }
-
-  private void removeExclusions(Map<String, CompiledClass> classes, List<String> exclusions) {
-    for (String exclusion : exclusions) {
-      exclusion = exclusion.replace('.', '/');
-      classes.remove(exclusion);
-    }
-  }
-
-  private List<ClassAndMethods> findToDatasAndFromDatas() {
-    List<ClassAndMethods> result = new ArrayList<>();
-    for (Map.Entry<String, CompiledClass> entry : this.classes.entrySet()) {
-      CompiledClass compiledClass = entry.getValue();
-      ClassAndMethods classAndMethods = null;
-
-      for (int i = 0; i < compiledClass.methods.length; i++) {
-        CompiledMethod method = compiledClass.methods[i];
-
-        if (!method.isAbstract() && method.descriptor().equals("void")) {
-          String name = method.name();
-          if (name.startsWith("toData") || name.startsWith("fromData")) {
-            if (classAndMethods == null) {
-              classAndMethods = new ClassAndMethods(compiledClass);
-            }
-            classAndMethods.methods.put(method.name(), method);
-          }
-        }
-      }
-      if (classAndMethods != null) {
-        result.add(classAndMethods);
-      }
-    }
-    Collections.sort(result);
-    return result;
-  }
 
   private List<ClassAndVariables> findSerializables() throws IOException {
     List<ClassAndVariables> result = new ArrayList<>(2000);
@@ -507,7 +261,7 @@ public abstract class AnalyzeSerializablesJUnitTestBase {
       }
       // System.out.println("processing class " + compiledClass.fullyQualifiedName());
 
-      if (!compiledClass.isInterface() && compiledClass.isSerializableAndNotDataSerializable()) {
+      if (!compiledClass.isInterface() && isSerializableAndNotDataSerializable(compiledClass)) {
         ClassAndVariables classAndVariables = new ClassAndVariables(compiledClass);
         for (int i = 0; i < compiledClass.fields_count; i++) {
           CompiledField compiledField = compiledClass.fields[i];
@@ -522,28 +276,83 @@ public abstract class AnalyzeSerializablesJUnitTestBase {
     return result;
   }
 
-  private File createEmptyFile(String fileName) throws IOException {
-    File file = new File(fileName);
-    if (file.exists()) {
-      assertThat(file.delete()).isTrue();
-    }
-    assertThat(file.createNewFile()).isTrue();
-    assertThat(file).exists().canWrite();
-    return file;
+
+
+  @Override
+  protected void initializeSerializationService() {
+    Properties properties = new Properties();
+    properties.put(ConfigurationProperties.VALIDATE_SERIALIZABLE_OBJECTS, "true");
+    properties.put(ConfigurationProperties.SERIALIZABLE_OBJECT_FILTER, "!*");
+    DistributionConfig distributionConfig = new DistributionConfigImpl(properties);
+    InternalDataSerializer.initialize(distributionConfig, initializeServices());
   }
 
-  /**
-   * Use this method to get a resource stored in the test's resource directory
-   */
-  private File getResourceAsFile(String resourceName) {
-    return new File(getClass().getResource(resourceName).getFile());
+  @Override
+  protected void deserializeObject(BufferDataOutputStream outputStream)
+      throws IOException, ClassNotFoundException {
+    DataSerializer
+        .readObject(new DataInputStream(new ByteArrayInputStream(outputStream.toByteArray())));
   }
 
-  /**
-   * Use this method to get a resource that might be in a JAR file
-   */
-  private InputStream getResourceAsStream(Class associatedClass, String resourceName)
+  @Override
+  protected void serializeObject(Object object, BufferDataOutputStream outputStream)
       throws IOException {
-    return associatedClass.getResource(resourceName).openStream();
+    DataSerializer.writeObject(object, outputStream);
   }
+
+
+  private void serializeAndDeserializeSanctionedObject(Object object) throws Exception {
+    BufferDataOutputStream outputStream = new BufferDataOutputStream(Version.CURRENT);
+    try {
+      serializeObject(object, outputStream);
+    } catch (IOException e) {
+      // some classes, such as BackupLock, are Serializable because the extend something
+      // like ReentrantLock but we never serialize them & it doesn't work to try to do so
+      throw new AssertionError("Not Serializable: " + object.getClass().getName(), e);
+    }
+    try {
+      deserializeObject(outputStream);
+    } catch (CancelException e) {
+      // PDX classes fish for a PDXRegistry and find that there is no cache
+    } catch (InvalidClassException e) {
+      fail("I was unable to deserialize " + object.getClass().getName(), e);
+    }
+  }
+
+  private List<DistributedSystemService> initializeServices() {
+    ServiceLoader<DistributedSystemService> loader =
+        ServiceLoader.load(DistributedSystemService.class);
+    List<DistributedSystemService> services = new ArrayList<>();
+    for (DistributedSystemService service : loader) {
+      services.add(service);
+    }
+    return services;
+  }
+
+
+  public boolean isSerializableAndNotDataSerializable(CompiledClass compiledClass) {
+    // these classes throw exceptions or log ugly messages when you try to load them
+    // in junit
+    String name = compiledClass.fullyQualifiedName().replace('/', '.');
+    if (name.startsWith("org.apache.geode.internal.shared.NativeCallsJNAImpl")
+        || name.startsWith("org.apache.geode.internal.statistics.HostStatHelper")) {
+      return false;
+    }
+    try {
+      Class realClass = Class.forName(name);
+      return Serializable.class.isAssignableFrom(realClass)
+          && !DataSerializable.class.isAssignableFrom(realClass)
+          && !DataSerializableFixedID.class.isAssignableFrom(realClass);
+    } catch (UnsatisfiedLinkError e) {
+      System.out.println("Unable to load actual class " + name + " external JNI dependencies");
+    } catch (NoClassDefFoundError e) {
+      System.out.println("Unable to load actual class " + name + " not in JUnit classpath");
+    } catch (Throwable e) {
+      System.out.println("Unable to load actual class " + name + ": " + e);
+    }
+    return false;
+  }
+
+
+
 }
