+++ geode-membership/src/integrationTest/resources/org/apache/geode/codeAnalysis/sanctionedDataSerializables.txt
@@ -1,0 +1,72 @@
+org/apache/geode/distributed/internal/membership/gms/GMSMembershipView,2
+fromData,121
+toData,72
+
+org/apache/geode/distributed/internal/membership/gms/MemberIdentifierImpl,6
+fromData,39
+fromDataPre_GFE_7_1_0_0,308
+fromDataPre_GFE_9_0_0_0,308
+toData,35
+toDataPre_GFE_7_1_0_0,290
+toDataPre_GFE_9_0_0_0,280
+
+org/apache/geode/distributed/internal/membership/gms/locator/FindCoordinatorRequest,2
+fromData,112
+toData,132
+
+org/apache/geode/distributed/internal/membership/gms/locator/FindCoordinatorResponse,2
+fromData,130
+toData,114
+
+org/apache/geode/distributed/internal/membership/gms/locator/GetViewRequest,2
+fromData,1
+toData,1
+
+org/apache/geode/distributed/internal/membership/gms/locator/GetViewResponse,2
+fromData,20
+toData,17
+
+org/apache/geode/distributed/internal/membership/gms/messages/FinalCheckPassedMessage,2
+fromData,20
+toData,17
+
+org/apache/geode/distributed/internal/membership/gms/messages/HeartbeatMessage,2
+fromData,11
+toData,11
+
+org/apache/geode/distributed/internal/membership/gms/messages/HeartbeatRequestMessage,2
+fromData,30
+toData,27
+
+org/apache/geode/distributed/internal/membership/gms/messages/InstallViewMessage,2
+fromData,60
+toData,56
+
+org/apache/geode/distributed/internal/membership/gms/messages/JoinRequestMessage,2
+fromData,63
+toData,60
+
+org/apache/geode/distributed/internal/membership/gms/messages/JoinResponseMessage,2
+fromData,63
+toData,57
+
+org/apache/geode/distributed/internal/membership/gms/messages/LeaveRequestMessage,2
+fromData,28
+toData,25
+
+org/apache/geode/distributed/internal/membership/gms/messages/NetworkPartitionMessage,2
+fromData,1
+toData,1
+
+org/apache/geode/distributed/internal/membership/gms/messages/RemoveMemberMessage,2
+fromData,28
+toData,25
+
+org/apache/geode/distributed/internal/membership/gms/messages/SuspectMembersMessage,2
+fromData,63
+toData,92
+
+org/apache/geode/distributed/internal/membership/gms/messages/ViewAckMessage,2
+fromData,40
+toData,37
+
